src/impl/generator/requirements.txt
docopt==0.6.1
scipy
numpy
pillow
av
nose
